Delete GradingSchemeTable component
